some speed profiling and parameter testing

+++ kuka_arm/src/stomp_trajectory_sampler.cpp
@@ -32,6 +32,7 @@ void setPoseTargetStomp(moveit::planning_interface::MoveGroupInterface &move_gro
   tf::poseMsgToEigen(pose, target_pose1);
   robot_state::RobotState robot_state = *move_group.getCurrentState();
   robot_state.setFromIK(joint_model_group, target_pose1, 5, 1.0);
+
   move_group.setJointValueTarget(robot_state);
 
 }
@@ -52,66 +53,41 @@ StompTrajectorySampler::StompTrajectorySampler(ros::NodeHandle nh)
   planning_scene::PlanningScene& planning_scene = *planning_scene_ptr;
   robot_state::RobotState robot_kinematic_state(kinematic_model);
 
+
   // set Stom as the planner and set allowed planning time
   move_group.setPlannerId("STOMP");
-  move_group.setPlanningTime(10.0);
+  move_group.setPlanningTime(100.0);
   eef_group.setPlannerId("STOMP");
-  eef_group.setPlanningTime(5.0);
+  eef_group.setPlanningTime(100.0);
 
   ROS_INFO("My Active planning scene: %s",planning_scene.getActiveCollisionDetectorName().c_str());
 
-  // Pointer to JointModelGroup for improved performance.
   joint_model_group =
       move_group.getCurrentState()->getJointModelGroup(PLANNING_GROUP);
   gripper_joint_model_group =
       eef_group.getCurrentState()->getJointModelGroup(GRIPPER_GROUP);
 
-  //addCollisionObjects();
 
-  while (ros::ok()) {
+  if(ros::ok()){
 
-    moveit_visual_tools::MoveItVisualTools visual_tools("world");
+    moveit_visual_tools::MoveItVisualTools visual_tools("base_footprint");
     visual_tools.deleteAllMarkers();
     visual_tools.loadRemoteControl();
 
 
-    Eigen::Affine3d text_pose = Eigen::Affine3d::Identity();
-    Eigen::Affine3d instr_pose = Eigen::Affine3d::Identity();
-    text_pose.translation().z() = 4.0;
-    instr_pose.translation().z() = 3.5;
-    visual_tools.publishText(text_pose, "Welcome to tsdf integration moveit",
-                             rviz_visual_tools::WHITE, rviz_visual_tools::XXXXLARGE);
-    visual_tools.trigger();
-    visual_tools.prompt("next step");
+    //Eigen::Affine3d text_pose = Eigen::Affine3d::Identity();
+    //Eigen::Affine3d instr_pose = Eigen::Affine3d::Identity();
+    //text_pose.translation().z() = 4.0;
+    //instr_pose.translation().z() = 3.5;
+
+    //visual_tools.publishText(text_pose, "Welcome to tsdf integration moveit",
+                             //rviz_visual_tools::WHITE, rviz_visual_tools::XXXXLARGE);
+    //visual_tools.trigger();
+    //visual_tools.prompt("next step");
 
 
-    float target_x, target_y, target_z;
-    float bin_x, bin_y, bin_z;
-
-    ros::param::get("/target_spawn_location/x", target_x);
-    ros::param::get("/target_spawn_location/y", target_y);
-    ros::param::get("/target_spawn_location/z", target_z);
-
-    ros::param::get("/target_drop_location/x", bin_x);
-    ros::param::get("/target_drop_location/y", bin_y);
-    ros::param::get("/target_drop_location/z", bin_z);
-
-    geometry_msgs::Pose target_pose, bin_pose, target_reach;
+    geometry_msgs::Pose target_pose;
     target_pose.orientation.w = 1.0;
-    target_pose.position.x = target_x - 0.4;
-    target_pose.position.y = target_y;
-    target_pose.position.z = target_z - 0.1;
-
-    target_reach.orientation.w = 1.0;
-    target_reach.position.x = target_x - 0.2;
-    target_reach.position.y = target_y;
-    target_reach.position.z = target_z - 0.1;
-
-    bin_pose.orientation.w = 1.0;
-    bin_pose.position.x = bin_x - 0.1;
-    bin_pose.position.y = bin_y;
-    bin_pose.position.z = bin_z + 1.6;
-
 
     setStartStateToCurrentStateStomp(move_group);
     setPoseTargetStomp(move_group, target_pose);
@@ -125,20 +101,20 @@ StompTrajectorySampler::StompTrajectorySampler(ros::NodeHandle nh)
     ROS_INFO("Visualizing plan to target: %s",
              success ? "SUCCEEDED" : "FAILED");
 
-    visualize_plan(visual_tools, text_pose, target_pose, my_plan);
-    visual_tools.trigger();
-    visual_tools.prompt("next step");
+    //visualize_plan(visual_tools, text_pose, target_pose, my_plan);
+    //visual_tools.trigger();
+    //visual_tools.prompt("next step");
 
     success = move_group.execute(my_plan) == moveit::planning_interface::MoveItErrorCode::SUCCESS;
     ROS_INFO("Moving to pick location: %s",
              success ? "SUCCEEDED" : "FAILED");
-
+    /*
     setStartStateToCurrentStateStomp(move_group);
     setPoseTargetStomp(move_group, target_reach);
 
     success = move_group.move() == moveit::planning_interface::MoveItErrorCode::SUCCESS;
     ROS_INFO("Target reach: %s",
-             success ? "SUCCEEDED" : "FAILED");
+             success ? "SUCCEEDED" : "FAILED");*/
   }
 }
 
@@ -151,68 +127,7 @@ void StompTrajectorySampler::visualize_plan(moveit_visual_tools::MoveItVisualToo
   visual_tools.publishTrajectoryLine(my_plan.trajectory_, joint_model_group);
 }
 
-void StompTrajectorySampler::addCollisionObjects() {/*
-   * Collision Objects:
-   * Create an object list and populate it with shelf and bin objects
-   * Then insert objects in scene for collision avoidance and interaction
-   */
-  std::vector<moveit_msgs::CollisionObject> collision_object_list;
-  moveit_msgs::CollisionObject shelf_collision_object, bin_collision_object,
-      ground_collision_object;
 
-  ground_collision_object.header.frame_id = move_group.getPlanningFrame();
-  ground_collision_object.id = "ground";
-
-  shape_msgs::SolidPrimitive ground_prim;
-  ground_prim.type = ground_prim.BOX;
-  ground_prim.dimensions.resize(3);
-  ground_prim.dimensions[0] = 6.0;
-  ground_prim.dimensions[1] = 6.0;
-  ground_prim.dimensions[2] = 0.05;
-
-  // Define pose for the objects (specified relative to base_footprint)
-  geometry_msgs::Pose bin_mesh_pose, shelf_mesh_pose, ground_prim_pose;
-  ground_prim_pose.orientation.w = 1.0;
-  ground_prim_pose.position.x = 0.0;
-  ground_prim_pose.position.y = 0.0;
-  ground_prim_pose.position.z = -0.025;
-
-  shelf_mesh_pose.position.x = 2.7;
-  shelf_mesh_pose.position.y = 0.0;
-  shelf_mesh_pose.position.z = 0.84;
-  shelf_mesh_pose.orientation.w = 0.707;
-  shelf_mesh_pose.orientation.x = 0.0;
-  shelf_mesh_pose.orientation.y = 0.0;
-  shelf_mesh_pose.orientation.z = 0.707;
-
-  bin_mesh_pose.position.x = 0.0;
-  bin_mesh_pose.position.y = 2.5;
-  bin_mesh_pose.position.z = 0.0;
-  bin_mesh_pose.orientation.w = 1.0;
-  bin_mesh_pose.orientation.x = 0.0;
-  bin_mesh_pose.orientation.y = 0.0;
-  bin_mesh_pose.orientation.z = 0.0;
-
-  SetupCollisionObject("shelf", SHELF_MESH_PATH, shelf_mesh_pose,
-                       shelf_collision_object);
-  SetupCollisionObject("bin", BIN_MESH_PATH, bin_mesh_pose,
-                       bin_collision_object);
-
-  ground_collision_object.primitives.push_back(ground_prim);
-  ground_collision_object.primitive_poses.push_back(ground_prim_pose);
-  ground_collision_object.operation = ground_collision_object.ADD;
-
-  collision_object_list.push_back(shelf_collision_object);
-  collision_object_list.push_back(bin_collision_object);
-  collision_object_list.push_back(ground_collision_object);
-
-  // Add the object list to the world scene
-  planning_scene_interface.addCollisionObjects(collision_object_list);
-  ROS_INFO("Added object list to the world");
-
-  // Allow MoveGroup to add the collision objects in the world
-  ros::Duration(1.0).sleep();
-}
 
 bool StompTrajectorySampler::OperateGripper(const bool &close_gripper) {
   // RobotState contains the current position/velocity/acceleration data
@@ -254,32 +169,7 @@ bool StompTrajectorySampler::CloseGripper() {
   return success;
 }
 
-bool StompTrajectorySampler::SetupCollisionObject(const std::string &object_id,
-                                             const std::string &mesh_path,
-                                             const geometry_msgs::Pose &object_pose,
-                                             moveit_msgs::CollisionObject &collision_object) {
-  collision_object.header.frame_id = move_group.getPlanningFrame();
-  collision_object.id = object_id;
 
-  shapes::Mesh *m = shapes::createMeshFromResource(mesh_path);
-
-  ROS_DEBUG_STREAM(object_id << " mesh loaded");
-
-  shape_msgs::Mesh object_mesh;
-  shapes::ShapeMsg object_mesh_msg;
-  shapes::constructMsgFromShape(m, object_mesh_msg);
-  object_mesh = boost::get<shape_msgs::Mesh>(object_mesh_msg);
-  collision_object.meshes.resize(1);
-  collision_object.mesh_poses.resize(1);
-  collision_object.meshes[0] = object_mesh;
-
-  collision_object.mesh_poses[0].position = object_pose.position;
-  collision_object.mesh_poses[0].orientation = object_pose.orientation;
-
-  collision_object.meshes.push_back(object_mesh);
-  collision_object.mesh_poses.push_back(collision_object.mesh_poses[0]);
-  collision_object.operation = collision_object.ADD;
-}
 
 
 StompTrajectorySampler::~StompTrajectorySampler() {}
